- remove geoxml_parameters_new_parameter as it is nonsense for group npar - fixed geoxml_sequence_append_clone to adjust npar - fixed adjust type

+++ libgebr/src/geoxml/geoxml/parameters_p.h
@@ -23,7 +23,7 @@
  * then adjust the number of parameters of a instance (npar)
  */
 gboolean
-__geoxml_parameters_adjust_group_npar(GeoXmlParameters * parameters, gint adjust);
+__geoxml_parameters_adjust_group_npar(GeoXmlParameters * parameters, glong adjust);
 
 /**
  * \internal
